加入資料結構基礎類型
