+++ cpp/100_same_tree.cpp
@@ -1,20 +1,7 @@
+#include "src/Tree.hpp"
 #include <iostream>
 
 using namespace std;
-
-struct TreeNode {
-  int val;
-  TreeNode *left;
-  TreeNode *right;
-
-  // Constructor.
-  TreeNode() : val(0), left(nullptr), right(nullptr) {}
-  TreeNode(int x) : val(x), left(nullptr), right(nullptr) {}
-  TreeNode(int x, TreeNode *left, TreeNode *right)
-      : val(x), left(left), right(right) {}
-
-  void bfs() {}
-};
 
 class Solution {
 public:
