+++ cpp/21_merge_two_sorted_lists.cpp
@@ -1,11 +1,4 @@
-struct ListNode {
-  int val;
-  ListNode *next;
-
-  ListNode() : val(0), next(nullptr) {}
-  ListNode(int x) : val(x), next(nullptr) {}
-  ListNode(int x, ListNode *next) : val(x), next(next) {}
-};
+#include "src/LinkedList.hpp"
 
 class Solution {
 public:
