+++ cpp/94_inorder_traversal.cpp
@@ -1,17 +1,8 @@
+#include "src/Tree.hpp"
 #include <iostream>
 #include <vector>
 
 using namespace std;
-
-struct TreeNode {
-  int val;
-  TreeNode *left;
-  TreeNode *right;
-
-  TreeNode() : val(0), left(nullptr), right(nullptr) {}
-  TreeNode(int x, ...) : val(x), left(nullptr), right(nullptr) {}
-  TreeNode(int x, TreeNode *l, TreeNode *r) : val(x), left(l), right(r) {}
-};
 
 class Solution {
 private:
