+++ cpp/src/Tree.hpp
@@ -1,0 +1,14 @@
+#ifndef TREE_HPP
+#define TREE_HPP
+
+struct TreeNode {
+  int val;
+  TreeNode *left;
+  TreeNode *right;
+
+  TreeNode() : val(0), left(nullptr), right(nullptr) {}
+  TreeNode(int x, ...) : val(x), left(nullptr), right(nullptr) {}
+  TreeNode(int x, TreeNode *l, TreeNode *r) : val(x), left(l), right(r) {}
+};
+
+#endif // !TREE_HPP
